Clicking on objects functionality
Very basic scaffolding for 'certain objects in the world can be clicked on to do something'. I like the method I'm using but it does require some kind of collider to function, which is likely fine, but worth remembering. I think more specific data on what happens when you click a given object probably can be handled with ScriptableObjects.

Also moved the drag-to-move-camera logic from PlayerMovement into CameraControl.