UI fiddling
A debug menu that you toggle with ` and a file to hold the code for it + other UI stuff.

I absolutely, positively need to switch to the new input system before I do too much else that hinges on input, because good lord I refuse to put all my input handling in a giant if-else tree that runs every frame. Assuming it's event-driven by default. If it's not I guess I have to write my own event handler hhrrgrglhl

Also, our current movement code registers the player as briefly airborne (it moves him up .05) when running into the chest for some reason, so that probably needs another pass. why is 'keep this guy on the floor' such a nightmare